refactor: fix code paths for code in some places
