+++ dagger-common/src/test/java/io/odpf/dagger/common/serde/deserialization/json/JsonDeserializerTest.java
@@ -1,0 +1,7 @@
+package io.odpf.dagger.common.serde.deserialization.json;
+
+import static org.junit.Assert.*;
+
+public class JsonDeserializerTest {
+
+}
