+++ dagger-core/build.gradle
@@ -59,10 +59,11 @@ configurations {
 // implementation: add all dependencies required just for local build
 // testImplementation: add all dependencies required just for tests
 dependencies {
-    compileOnly 'org.projectlombok:lombok:1.18.8'
-    annotationProcessor 'org.projectlombok:lombok:1.18.8'
     minimalJar project(path: ':dagger-common', configuration: 'minimalCommonJar')
     minimalJar project(path: ':dagger-functions', configuration: 'minimalFunctionsJar')
+
+    compileOnly 'org.projectlombok:lombok:1.18.8'
+    annotationProcessor 'org.projectlombok:lombok:1.18.8'
 
     implementation 'org.slf4j:slf4j-log4j12:1.7.7'
     implementation 'org.apache.flink:flink-streaming-java_2.11:' + flinkVersion
