+++ dagger-core/src/test/java/io/odpf/dagger/core/sink/kafka/KafkaSerializationSchemaFactoryTest.java
@@ -1,0 +1,7 @@
+package io.odpf.dagger.core.sink.kafka;
+
+import static org.junit.Assert.*;
+
+public class KafkaSerializationSchemaFactoryTest {
+
+}
