+++ dagger-core/src/test/java/io/odpf/dagger/core/sink/kafka/builder/KafkaJsonSerializerBuilderTest.java
@@ -1,0 +1,7 @@
+package io.odpf.dagger.core.sink.kafka.builder;
+
+import static org.junit.Assert.*;
+
+public class KafkaJsonSerializerBuilderTest {
+
+}
