+++ dagger-core/src/test/java/io/odpf/dagger/core/sink/kafka/builder/KafkaProtoSerializerBuilderTest.java
@@ -1,0 +1,22 @@
+package io.odpf.dagger.core.sink.kafka.builder;
+
+import org.junit.Before;
+import org.junit.Test;
+
+public class KafkaProtoSerializerBuilderTest {
+
+    @Before
+    public void setup() {
+
+    }
+
+    @Test
+    public void shouldCreateProtoSerializer() {
+
+    }
+
+    @Test
+    public void shouldAddMetrics() {
+
+    }
+}
