+++ dagger-core/src/test/java/io/odpf/dagger/core/stream/StreamsFactoryTest.java
@@ -1,0 +1,7 @@
+package io.odpf.dagger.core.stream;
+
+import static org.junit.Assert.*;
+
+public class StreamsFactoryTest {
+
+}
