+++ dagger-core/src/test/java/io/odpf/dagger/core/stream/builder/JsonDataStreamBuilderTest.java
@@ -1,0 +1,14 @@
+package io.odpf.dagger.core.stream.builder;
+
+import org.junit.Test;
+
+public class JsonDataStreamBuilderTest {
+
+
+
+    // TODO : this will be bit tricky class to handle
+    @Test
+    public void shouldReturnProducedType() {
+
+    }
+}
