+++ dagger-core/src/test/java/io/odpf/dagger/core/stream/builder/ProtoDataStreamBuilderTest.java
@@ -1,0 +1,7 @@
+package io.odpf.dagger.core.stream.builder;
+
+import static org.junit.Assert.*;
+
+public class ProtoDataStreamBuilderTest {
+
+}
